+++ tests/test_xkcd.py
@@ -1,4 +1,4 @@
-from Dev.Chromaspace.src.Chromaspace.xkcd import rgb_distance
+from Chromaspace.xkcd import rgb_distance
 
 def test_rgb_distance():
     assert round(rgb_distance([0,0,0], [0,0,0]), 6) == 0